docs: refactor logo
