+++ packages/dnb-design-system-portal/src/pages/uilib/components/logo.md
@@ -5,11 +5,18 @@ status: null
 order: 10
 ---
 
-<!--
-  ATTENTION: This file is auto generated by using "makeDemosFactory".
-  Do not change the content!
--->
+import Tabs from 'Tags/Tabs'
 
-import LogoPage from 'Src/uilib/components/demos/Logo'
+import LogoInfo from 'Pages/uilib/components/logo/logo-info'
+import LogoProperties from 'Pages/uilib/components/logo/logo-properties'
 
-<LogoPage />
+# Logo
+
+<Tabs>
+  <Tabs.Content>
+    <LogoInfo />
+  </Tabs.Content>
+  <Tabs.Content>
+    <LogoProperties />
+  </Tabs.Content>
+</Tabs>
