+++ packages/dnb-design-system-portal/src/pages/uilib/components/logo/Examples.js
@@ -1,0 +1,44 @@
+/**
+ * UI lib Component Example
+ *
+ */
+
+import React, { PureComponent } from 'react'
+import ComponentBox from '../../../../shared/tags/ComponentBox'
+
+class Example extends PureComponent {
+  render() {
+    return (
+      <>
+        <ComponentBox
+          caption="DNB Logo with dynamic height"
+          data-dnb-test="logo-auto-size"
+        >
+          {/* @jsx */ `
+<span style={{fontSize: '12rem'}}>
+  <Logo size="auto" />
+</span>
+          `}
+        </ComponentBox>
+        <ComponentBox
+          caption="DNB Logo with dynamic height"
+          data-dnb-test="logo-inherit-size"
+        >
+          {/* @jsx */ `
+<span style={{height: '12rem'}}>
+  <Logo size="inherit" />
+</span>
+          `}
+        </ComponentBox>
+        <ComponentBox caption="DNB Logo with fixed height">
+          {/* @jsx */ `
+<Logo height="192" data-dnb-test="logo-default" />
+        `}
+        </ComponentBox>
+      </>
+    )
+  }
+}
+
+export { Example }
+export default () => <Example />
