+++ packages/dnb-design-system-portal/src/pages/uilib/components/logo/logo-info.md
@@ -1,0 +1,13 @@
+---
+draft: true
+---
+
+import Examples from 'Pages/uilib/components/logo/Examples'
+
+## Description
+
+A ready to use DNB logo in SVG format.
+
+## Demos
+
+<Examples />
