--- packages/dnb-design-system-portal/src/uilib/components/examples/Logo.txt
@@ -1,1 +1,0 @@
-<Logo height="200" />
