+++ packages/dnb-ui-lib/src/components/logo/__tests__/Logo.screenshot.test.js
@@ -16,4 +16,16 @@ describe('Logo screenshot', () => {
     })
     expect(screenshot).toMatchImageSnapshot()
   })
+  it('have to match the auto sized "Logo"', async () => {
+    const screenshot = await testPageScreenshot({
+      selector: '[data-dnb-test="logo-auto-size"]'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
+  it('have to match the inherited sized "Logo"', async () => {
+    const screenshot = await testPageScreenshot({
+      selector: '[data-dnb-test="logo-inherit-size"]'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
 })
