--- packages/dnb-ui-lib/src/components/logo/description.md
@@ -1,3 +1,0 @@
-## Description
-
-A ready to use DNB logo in SVG format.
